Move Issue30 tests into SharedTagTest as requested by maintainer
